+++ src/types/query/loyaltyProgram.ts
@@ -1,0 +1,5 @@
+export type LoyaltyProgramInfo = {
+  pendingReward: number;
+  totalPoolReward: number;
+  isClaimEnabled: boolean;
+};