+++ src/utils/builder.ts
@@ -22,16 +22,61 @@ export const requireSender = (txBlock: SuiKitTxBlock) => {
   return sender;
 };
 
+export const checkVesca = (prevUnlockAtInMillisTimestamp?: number) => {
+  if (prevUnlockAtInMillisTimestamp === undefined) {
+    throw new Error('veSca not found');
+  }
+};
+
+export const checkVescaExpired = (prevUnlockAtInMillisTimestamp: number) => {
+  if (prevUnlockAtInMillisTimestamp <= new Date().getTime()) {
+    throw new Error('veSca is expired, use renewExpiredVeScaQuick instead');
+  }
+};
+
+export const checkExtendLockPeriod = (
+  lockPeriodInDays: number,
+  newUnlockAtInSecondTimestamp: number,
+  prevUnlockAtInMillisTimestamp?: number
+) => {
+  checkVesca(prevUnlockAtInMillisTimestamp);
+  checkVescaExpired(prevUnlockAtInMillisTimestamp!);
+  const prevUnlockAtInSecondTimestamp = Math.floor(
+    prevUnlockAtInMillisTimestamp! / 1000
+  );
+  if (lockPeriodInDays < 1) {
+    throw new Error('Minimum lock period is 1 day');
+  }
+
+  const availableLockPeriodInDays = Math.floor(
+    (newUnlockAtInSecondTimestamp - prevUnlockAtInSecondTimestamp) /
+      UNLOCK_ROUND_DURATION
+  );
+  console.log('availableLockPeriodInDays', availableLockPeriodInDays);
+  if (lockPeriodInDays > availableLockPeriodInDays) {
+    throw new Error(
+      `Cannot extend lock period by ${lockPeriodInDays} days, maximum lock period is ~4 years (${MAX_LOCK_ROUNDS} days), remaining lock period is ${
+        MAX_LOCK_ROUNDS - availableLockPeriodInDays
+      }`
+    );
+  }
+};
+
 export const checkLockSca = (
-  scaAmountOrCoin?: number | SuiObjectArg | undefined,
+  scaAmountOrCoin: number | SuiObjectArg | undefined,
   lockPeriodInDays?: number,
   newUnlockAtInSecondTimestamp?: number,
-  prevUnlockAtInSecondTimestamp?: number
+  prevUnlockAtInMillisTimestamp?: number
 ) => {
+  const prevUnlockAtInSecondTimestamp = prevUnlockAtInMillisTimestamp
+    ? Math.floor(prevUnlockAtInMillisTimestamp / 1000)
+    : undefined;
   const isInitialLock = !prevUnlockAtInSecondTimestamp;
   const isLockExpired =
     !isInitialLock &&
     prevUnlockAtInSecondTimestamp * 1000 <= new Date().getTime();
+
+  // handle for initial lock / renewing expired veSca
   if (isInitialLock || isLockExpired) {
     if (scaAmountOrCoin !== undefined && lockPeriodInDays !== undefined) {
       if (lockPeriodInDays <= 0) {
@@ -61,7 +106,9 @@ export const checkLockSca = (
       );
     }
   } else {
-    checkVesca(prevUnlockAtInSecondTimestamp);
+    // handle for extending lock period / top up / both
+    checkVesca(prevUnlockAtInMillisTimestamp);
+    checkVescaExpired(prevUnlockAtInMillisTimestamp!);
     if (
       typeof scaAmountOrCoin === 'number' &&
       scaAmountOrCoin < MIN_TOP_UP_AMOUNT
@@ -69,43 +116,12 @@ export const checkLockSca = (
       throw new Error('Minimum top up amount is 1 SCA');
     }
 
-    if (!!newUnlockAtInSecondTimestamp && !!prevUnlockAtInSecondTimestamp) {
-      const totalLockDuration =
-        newUnlockAtInSecondTimestamp - prevUnlockAtInSecondTimestamp;
-      if (totalLockDuration > MAX_LOCK_DURATION - UNLOCK_ROUND_DURATION) {
-        throw new Error(
-          `Maximum lock period is ~4 years (${MAX_LOCK_ROUNDS - 1} days)`
-        );
-      }
-    }
-  }
-};
-
-export const checkExtendLockPeriod = (
-  lockPeriodInDays: number,
-  newUnlockAtInSecondTimestamp: number,
-  prevUnlockAtInSecondTimestamp?: number
-) => {
-  checkVesca(prevUnlockAtInSecondTimestamp);
-
-  if (lockPeriodInDays <= 0) {
-    throw new Error('Lock period must be greater than 0');
-  }
-
-  const isInitialLock = !prevUnlockAtInSecondTimestamp;
-  const isLockExpired =
-    !isInitialLock &&
-    prevUnlockAtInSecondTimestamp * 1000 <= new Date().getTime();
-  if (isLockExpired) {
-    throw new Error('veSca is expired, use renewExpiredVeScaQuick instead');
-  }
-
-  if (prevUnlockAtInSecondTimestamp) {
-    const totalLockDuration =
-      newUnlockAtInSecondTimestamp - prevUnlockAtInSecondTimestamp!;
-    if (totalLockDuration > MAX_LOCK_DURATION - UNLOCK_ROUND_DURATION) {
-      throw new Error(
-        `Maximum lock period is ~4 years (${MAX_LOCK_ROUNDS - 1} days)`
+    // for topup and extend lock period
+    if (newUnlockAtInSecondTimestamp && lockPeriodInDays) {
+      checkExtendLockPeriod(
+        lockPeriodInDays,
+        newUnlockAtInSecondTimestamp,
+        prevUnlockAtInMillisTimestamp
       );
     }
   }
@@ -113,18 +129,19 @@ export const checkExtendLockPeriod = (
 
 export const checkExtendLockAmount = (
   scaAmount: number,
-  prevUnlockAtInSecondTimestamp?: number
+  prevUnlockAtInMillisTimestamp?: number
 ) => {
-  checkVesca(prevUnlockAtInSecondTimestamp);
+  checkVesca(prevUnlockAtInMillisTimestamp);
+  checkVescaExpired(prevUnlockAtInMillisTimestamp!);
 
   if (scaAmount < MIN_TOP_UP_AMOUNT) {
     throw new Error('Minimum top up amount is 1 SCA');
   }
 
-  const isInitialLock = !prevUnlockAtInSecondTimestamp;
+  const isInitialLock = !prevUnlockAtInMillisTimestamp;
   const isLockExpired =
-    !isInitialLock &&
-    prevUnlockAtInSecondTimestamp * 1000 <= new Date().getTime();
+    !isInitialLock && prevUnlockAtInMillisTimestamp <= new Date().getTime();
+
   if (isLockExpired) {
     throw new Error('veSca is expired, use renewExpiredVeScaQuick instead');
   }
@@ -133,9 +150,14 @@ export const checkExtendLockAmount = (
 export const checkRenewExpiredVeSca = (
   scaAmount: number,
   lockPeriodInDays: number,
-  prevUnlockAtInSecondTimestamp?: number
+  prevUnlockAtInMillisTimestamp?: number
 ) => {
-  checkVesca(prevUnlockAtInSecondTimestamp);
+  if (
+    !prevUnlockAtInMillisTimestamp ||
+    prevUnlockAtInMillisTimestamp > new Date().getTime()
+  ) {
+    throw new Error('Renew method can only be used for expired veSca');
+  }
 
   if (scaAmount < MIN_INITIAL_LOCK_AMOUNT) {
     throw new Error('Minimum lock amount for renewing expired vesca 10 SCA');
@@ -146,11 +168,5 @@ export const checkRenewExpiredVeSca = (
     throw new Error(
       `Maximum lock period is ~4 years (${MAX_LOCK_ROUNDS - 1} days)`
     );
-  }
-};
-
-export const checkVesca = (prevUnlockAtInSecondTimestamp?: number) => {
-  if (prevUnlockAtInSecondTimestamp === undefined) {
-    throw new Error('veSca not found');
   }
 };